+++ example/lib/index.dart
@@ -1,0 +1,4 @@
+export 'package:crypto_simple/crypto_simple.dart';
+export 'package:flutter/material.dart';
+
+export 'customWidgets.dart';