+++ lib/index.dart
@@ -1,0 +1,11 @@
+library crypto_simple;
+
+export 'dart:convert';
+export 'dart:io';
+export 'dart:math';
+export 'dart:typed_data';
+export 'package:crypto_simple/crypto_simple.dart';
+export 'package:flutter/foundation.dart';
+export 'package:flutter/material.dart';
+
+export 'crypto_simple.dart';