+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+* add Randomized mode for encryption/decryption .
+* add XOR method for encryption/decryption . 
+* add SUPER_XOR method for encryption/decryption .
+
 ## 1.1.1
 
 *  add subKey on constructor.